refactor: emulator 패키지 구조 바꾸기
