+++ back/src/main/java/com/example/_thecore_back/emulator/controller/dto/EmulatorConverter.java
@@ -1,6 +1,6 @@
-package com.example._thecore_back.rest.emulator.model;
+package com.example._thecore_back.emulator.controller.dto;
 
-import com.example._thecore_back.rest.emulator.db.EmulatorEntity;
+import com.example._thecore_back.emulator.domain.EmulatorEntity;
 import org.springframework.stereotype.Component;
 
 @Component
