+++ back/src/main/java/com/example/_thecore_back/emulator/controller/dto/FinalGetAllEmulatorResponse.java
@@ -1,4 +1,4 @@
-package com.example._thecore_back.rest.emulator.model;
+package com.example._thecore_back.emulator.controller.dto;
 
 import lombok.Builder;
 import lombok.Getter;
