+++ back/src/main/java/com/example/_thecore_back/emulator/controller/dto/GetEmulatorResponseData.java
@@ -1,11 +1,12 @@
-package com.example._thecore_back.rest.emulator.model;
+package com.example._thecore_back.emulator.controller.dto;
 
+import com.example._thecore_back.emulator.domain.EmulatorStatus;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
-public class RegisterEmulatorResponseData {
+public class GetEmulatorResponseData {
     private Long emulatorId;
     private String carNumber;
     private EmulatorStatus emulatorStatus;
