+++ back/src/main/java/com/example/_thecore_back/emulator/controller/dto/RegisterEmulatorResponseData.java
@@ -1,11 +1,12 @@
-package com.example._thecore_back.rest.emulator.model;
+package com.example._thecore_back.emulator.controller.dto;
 
+import com.example._thecore_back.emulator.domain.EmulatorStatus;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
-public class GetEmulatorResponseData {
+public class RegisterEmulatorResponseData {
     private Long emulatorId;
     private String carNumber;
     private EmulatorStatus emulatorStatus;
