+++ back/src/main/java/com/example/_thecore_back/emulator/domain/EmulatorStatus.java
@@ -1,0 +1,6 @@
+package com.example._thecore_back.emulator.domain;
+
+public enum EmulatorStatus {
+    ON,
+    OFF
+}
