+++ back/src/main/java/com/example/_thecore_back/emulator/domain/EmulatorWriter.java
@@ -1,0 +1,6 @@
+package com.example._thecore_back.emulator.domain;
+
+public interface EmulatorWriter {
+    EmulatorEntity save(EmulatorEntity emulator);
+    void delete(EmulatorEntity emulator);
+}
