+++ back/src/main/java/com/example/_thecore_back/emulator/infrastructure/EmulatorRepository.java
@@ -1,5 +1,6 @@
-package com.example._thecore_back.rest.emulator.db;
+package com.example._thecore_back.emulator.infrastructure;
 
+import com.example._thecore_back.emulator.domain.EmulatorEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
