--- back/src/main/java/com/example/_thecore_back/rest/emulator/model/EmulatorStatus.java
@@ -1,6 +1,0 @@
-package com.example._thecore_back.rest.emulator.model;
-
-public enum EmulatorStatus {
-    ON,
-    OFF
-}
